Introduce RenderService
Used to be RenderCoordinator
