+++ src/renderer/SelectionRenderLayer.ts
@@ -4,7 +4,7 @@
  */
 
 import { ITerminal } from '../Types';
-import { IRenderDimensions } from './Types';
+import { IRenderDimensions } from 'browser/renderer/Types';
 import { BaseRenderLayer } from './BaseRenderLayer';
 import { IColorSet } from 'browser/Types';
