+++ src/renderer/Types.ts
@@ -3,9 +3,10 @@
  * @license MIT
  */
 
-import { ITerminal, CharacterJoinerHandler } from '../Types';
+import { ITerminal } from '../Types';
 import { IDisposable } from 'xterm';
 import { IColorSet } from 'browser/Types';
+import { IRenderDimensions, CharacterJoinerHandler } from 'browser/renderer/Types';
 
 /**
  * Flags used to render terminal text properly.
@@ -18,44 +19,6 @@ export const enum FLAGS {
   INVISIBLE = 16,
   DIM = 32,
   ITALIC = 64
-}
-
-/**
- * Note that IRenderer implementations should emit the refresh event after
- * rendering rows to the screen.
- */
-export interface IRenderer extends IDisposable {
-  readonly dimensions: IRenderDimensions;
-
-  dispose(): void;
-  setColors(colors: IColorSet): void;
-  onDevicePixelRatioChange(): void;
-  onResize(cols: number, rows: number): void;
-  onCharSizeChanged(): void;
-  onBlur(): void;
-  onFocus(): void;
-  onSelectionChanged(start: [number, number], end: [number, number], columnSelectMode: boolean): void;
-  onCursorMove(): void;
-  onOptionsChanged(): void;
-  clear(): void;
-  renderRows(start: number, end: number): void;
-  registerCharacterJoiner(handler: CharacterJoinerHandler): number;
-  deregisterCharacterJoiner(joinerId: number): boolean;
-}
-
-export interface IRenderDimensions {
-  scaledCharWidth: number;
-  scaledCharHeight: number;
-  scaledCellWidth: number;
-  scaledCellHeight: number;
-  scaledCharLeft: number;
-  scaledCharTop: number;
-  scaledCanvasWidth: number;
-  scaledCanvasHeight: number;
-  canvasWidth: number;
-  canvasHeight: number;
-  actualCellWidth: number;
-  actualCellHeight: number;
 }
 
 export interface IRenderLayer extends IDisposable {
