+++ src/renderer/dom/DomRenderer.ts
@@ -3,8 +3,8 @@
  * @license MIT
  */
 
-import { IRenderer, IRenderDimensions } from '../Types';
-import { ILinkifierEvent, ITerminal, CharacterJoinerHandler } from '../../Types';
+import { IRenderer, IRenderDimensions, CharacterJoinerHandler } from 'browser/renderer/Types';
+import { ILinkifierEvent, ITerminal } from '../../Types';
 import { BOLD_CLASS, ITALIC_CLASS, CURSOR_CLASS, CURSOR_STYLE_BLOCK_CLASS, CURSOR_BLINK_CLASS, CURSOR_STYLE_BAR_CLASS, CURSOR_STYLE_UNDERLINE_CLASS, DomRendererRowFactory } from './DomRendererRowFactory';
 import { INVERTED_DEFAULT_COLOR } from '../atlas/Types';
 import { Disposable } from 'common/Lifecycle';
